2011-07-14 - Rev 8133 =====================
  :Author: Heino Falcke <H.Falcke@astro.ru.nl>

- added :func:`trerun` to store and rerun task instances (like trun)
  without recreating them 

-  task averagespectrum can now also form an incoherent sum of time
   series data for all antennas, fixed some bugs

- cr_event pipeline further developed, more graphs added, fixed
   delay fitting, normalized to Galactic powerlaw shape

- added functions hPowerlawMul, hPowerlawAdd, hLinearFunctionMul,
  hLinearFunctionAdd, hLogLinearFunctionMul, hLogLinearFunctionAdd



git-svn-id: http://usg.lofar.org/svn/code/trunk@8133 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/Pypeline/modules/tasks/averagespectrum.py
@@ -373,7 +373,7 @@ class WorkSpace(tasks.WorkSpace(taskname="AverageSpectrum")):
                      "Sample length in raw data set.",
                      "s"),
 
-        "blocksize":p_(lambda ws:min(1./ws.delta_nu/ws.delta_t,ws.datafile["DATA_LENGTH"][0]),"The desired blocksize for a single FFT. Can be set directly, otherwise estimated from delta_nu. The actual size will be rounded to give a power of two, hence, see ``blocksize_used` for the actual blocksize used."),
+        "blocksize":p_(lambda ws:int(min(1./ws.delta_nu/ws.delta_t,ws.datafile["DATA_LENGTH"][0])),"The desired blocksize for a single FFT. Can be set directly, otherwise estimated from delta_nu. The actual size will be rounded to give a power of two, hence, see ``blocksize_used` for the actual blocksize used."),
         
         "blocksize_used":p_(lambda ws:ws.blocklen*ws.nblocks,"The blocksize one would use for a single FFT."),
